requirements.txt
asgiref==3.6.0
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==1.3.13
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2018.5
sqlparse==0.4.1
tushare==1.2.48
